# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: KMRL Train Induction System API
  description: |
    AI-driven train scheduling and optimization system for Kochi Metro Rail Limited (KMRL).
    
    This comprehensive API provides endpoints for managing trainsets, fitness certificates, 
    job cards, schedules, analytics, and user management with real-time WebSocket support.
    
    ## Features
    - **Trainset Management**: Complete CRUD operations for metro trainsets
    - **Fitness Certificates**: Track and manage safety compliance certificates
    - **Job Cards**: Integration with IBM Maximo for maintenance workflows
    - **AI Optimization**: Advanced scheduling algorithms with machine learning
    - **Real-time Updates**: WebSocket support for live data streaming
    - **Analytics**: Comprehensive reporting and performance metrics
    - **Security**: JWT-based authentication with role-based access control
    
    ## Authentication
    This API uses JWT tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
  version: 1.0.0
  contact:
    name: KMRL Development Team
    email: tech@kmrl.gov.in
    url: https://kochimetro.org
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.kmrl-train-system.gov.in/api
    description: Production server
...
paths:
...
  /job-cards:
    get:
      tags: [Job Cards]
      summary: Get job cards
      description: Retrieve paginated list of job cards
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/JobStatus'
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/JobPriority'
        - name: overdue
          in: query
          schema:
            type: boolean
          description: Filter overdue job cards
      responses:
        '200':
          description: Job cards retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      jobCards:
                        type: array
                        items:
                          $ref: '#/components/schemas/JobCard'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  /job-cards/sync/maximo:
    post:
      tags: [Job Cards]
      summary: Sync with Maximo
      description: Synchronize job cards with IBM Maximo system
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Maximo sync completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Maximo sync completed
                  data:
                    type: object
                    properties:
                      syncedCount:
                        type: integer
                        example: 5
                      errors:
                        type: array
                        items:
                          type: string
                      timestamp:
                        type: string
                        format: date-time
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...
    JobCard:
      type: object
      properties:
        id:
          type: string
        trainsetId:
          type: string
        jobCardNumber:
          type: string
        maximoId:
          type: string
        priority:
          $ref: '#/components/schemas/JobPriority'
        status:
          $ref: '#/components/schemas/JobStatus'
        workType:
          type: string
        description:
          type: string
        estimatedHours:
          type: number
        actualHours:
          type: number
        assignedTo:
          type: string
        scheduledDate:
          type: string
          format: date-time
        completedDate:
          type: string
          format: date-time
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
    JobStatus:
      type: string
      enum:
        - PENDING
        - IN_PROGRESS
        - COMPLETED
        - CANCELLED
        - ON_HOLD
    JobPriority:
      type: string
      enum:
        - LOW
        - MEDIUM
        - HIGH
        - CRITICAL
        - EMERGENCY
...